fix: collect pb used items
